+++ src/write_outputs/write_capacity.jl
@@ -26,6 +26,11 @@ function write_capacity(path::AbstractString, inputs::Dict, setup::Dict, EP::Mod
 		end
 	end
 
+	capacity_constraint_dual = fill(NaN, size(inputs["RESOURCES"]))
+	for y in dfGen[dfGen.Max_Cap_MW.>0, :R_ID]
+		capacity_constraint_dual[y] = dual.(EP[:cMaxCap][y])
+	end
+
 	capcharge = zeros(size(inputs["RESOURCES"]))
 	retcapcharge = zeros(size(inputs["RESOURCES"]))
 	existingcapcharge = zeros(size(inputs["RESOURCES"]))
@@ -57,6 +62,7 @@ function write_capacity(path::AbstractString, inputs::Dict, setup::Dict, EP::Mod
 		RetCap = retcapdischarge[:],
 		NewCap = capdischarge[:],
 		EndCap = value.(EP[:eTotalCap]),
+		CapacityConstraintDual = capacity_constraint_dual[:],
 		StartEnergyCap = existingcapenergy[:],
 		RetEnergyCap = retcapenergy[:],
 		NewEnergyCap = capenergy[:],
@@ -71,6 +77,7 @@ function write_capacity(path::AbstractString, inputs::Dict, setup::Dict, EP::Mod
 		dfCap.RetCap = dfCap.RetCap * ModelScalingFactor
 		dfCap.NewCap = dfCap.NewCap * ModelScalingFactor
 		dfCap.EndCap = dfCap.EndCap * ModelScalingFactor
+		dfCap.CapacityConstraintDual = dfCap.CapacityConstraintDual * ModelScalingFactor
 		dfCap.StartEnergyCap = dfCap.StartEnergyCap * ModelScalingFactor
 		dfCap.RetEnergyCap = dfCap.RetEnergyCap * ModelScalingFactor
 		dfCap.NewEnergyCap = dfCap.NewEnergyCap * ModelScalingFactor
@@ -84,6 +91,7 @@ function write_capacity(path::AbstractString, inputs::Dict, setup::Dict, EP::Mod
 			Resource = "Total", Zone = "n/a",
 			StartCap = sum(dfCap[!,:StartCap]), RetCap = sum(dfCap[!,:RetCap]),
 			NewCap = sum(dfCap[!,:NewCap]), EndCap = sum(dfCap[!,:EndCap]),
+			CapacityConstraintDual = NaN,
 			StartEnergyCap = sum(dfCap[!,:StartEnergyCap]), RetEnergyCap = sum(dfCap[!,:RetEnergyCap]),
 			NewEnergyCap = sum(dfCap[!,:NewEnergyCap]), EndEnergyCap = sum(dfCap[!,:EndEnergyCap]),
 			StartChargeCap = sum(dfCap[!,:StartChargeCap]), RetChargeCap = sum(dfCap[!,:RetChargeCap]),