+++ CHANGELOG.md
@@ -23,6 +23,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Eliminate 'Axis contains one element' warning (seen when LDS is used) by combining two constraints (#496).
 
+### Added
+- Add output for dual of capacity constraint (#473)
+
 ## [0.3.5] - 2023-05-18
 
 ### Added